+++ lib/punchblock/component/record.rb
@@ -9,11 +9,14 @@ module Punchblock
       # Creates an Rayo Record command
       #
       # @param [Hash] options
-      # @option options [String, Optional] :text to speak back
-      # @option options [String, Optional] :voice with which to render TTS
-      # @option options [String, Optional] :ssml document to render TTS
+      # @option options [String] :format to use for recording
+      # @option options [Integer] :initial_timeout Controls how long the recognizer should wait after the end of the prompt for the caller to speak before sending a Recorder event.
+      # @option options [Integer] :final_timeout Controls the length of a period of silence after callers have spoken to conclude they finished.
+      # @option options [Integer] :max_duration Indicates the maximum duration for the recording.
+      # @option options [true, false] :start_beep Indicates whether subsequent record will be preceded with a beep.
+      # @option options [true, false] :start_paused Whether subsequent record will start in PAUSE mode.
       #
-      # @return [Command::Record] an Rayo "record" command
+      # @return [Command::Record] a Rayo "record" command
       #
       # @example
       #   record :text => 'Hello brown cow.'
@@ -28,14 +31,14 @@ module Punchblock
       end
 
       ##
-      # @return [String] the codec to use for recording
+      # @return [Integer] Controls the length of a period of silence after callers have spoken to conclude they finished.
       #
       def final_timeout
         read_attr :'final-timeout', :to_i
       end
 
       ##
-      # @param [String] codec to use for recording
+      # @param [Integer] timeout Controls the length of a period of silence after callers have spoken to conclude they finished.
       #
       def final_timeout=(timeout)
         write_attr :'final-timeout', timeout
@@ -56,77 +59,63 @@ module Punchblock
       end
 
       ##
-      # @return [String] the codec to use for recording
+      # @return [Integer] Controls how long the recognizer should wait after the end of the prompt for the caller to speak before sending a Recorder event.
       #
       def initial_timeout
         read_attr :'initial-timeout', :to_i
       end
 
       ##
-      # @param [String] codec to use for recording
+      # @param [Integer] timeout Controls how long the recognizer should wait after the end of the prompt for the caller to speak before sending a Recorder event.
       #
       def initial_timeout=(timeout)
         write_attr :'initial-timeout', timeout
       end
 
       ##
-      # @return [String] the codec to use for recording
+      # @return [Integer] Indicates the maximum duration for the recording.
       #
       def max_duration
         read_attr :'max-duration', :to_i
       end
 
       ##
-      # @param [String] codec to use for recording
+      # @param [Integer] other Indicates the maximum duration for the recording.
       #
       def max_duration=(other)
         write_attr :'max-duration', other
       end
 
       ##
-      # @return [String] the codec to use for recording
+      # @return [true, false] Indicates whether subsequent record will be preceded with a beep.
       #
       def start_beep
         read_attr(:'start-beep') == 'true'
       end
 
       ##
-      # @param [String] codec to use for recording
+      # @param [true, false] sb Indicates whether subsequent record will be preceded with a beep.
       #
       def start_beep=(sb)
         write_attr :'start-beep', sb
       end
 
       ##
-      # @return [String] the codec to use for recording
-      #
-      def stop_beep
-        read_attr(:'stop-beep') == 'true'
-      end
-
-      ##
-      # @param [String] codec to use for recording
-      #
-      def stop_beep=(sb)
-        write_attr :'stop-beep', sb
-      end
-
-      ##
-      # @return [String] the codec to use for recording
+      # @return [true, false] Whether subsequent record will start in PAUSE mode.
       #
       def start_paused
         read_attr(:'start-paused') == 'true'
       end
 
       ##
-      # @param [String] codec to use for recording
+      # @param [true, false] other Whether subsequent record will start in PAUSE mode.
       #
       def start_paused=(other)
         write_attr :'start-paused', other
       end
 
       def inspect_attributes # :nodoc:
-        [:final_timeout, :format, :initial_timeout, :max_duration, :start_beep, :start_paused, :stop_beep] + super
+        [:final_timeout, :format, :initial_timeout, :max_duration, :start_beep, :start_paused] + super
       end
 
       state_machine :state do
@@ -191,15 +180,15 @@ module Punchblock
       ##
       # Directly returns the recording for the component
       # @return [Punchblock::Component::Record::Recording] The recording object
-      # 
+      #
       def recording
         complete_event.recording
       end
-      
+
       ##
       # Directly returns the recording URI for the component
       # @return [String] The recording URI
-      # 
+      #
       def recording_uri
         recording.uri
       end