+++ lib/punchblock/event/complete.rb
@@ -3,10 +3,6 @@
 module Punchblock
   class Event
     class Complete < Event
-      # TODO: Validate response and return response type.
-      # -----
-      # <complete xmlns="urn:xmpp:rayo:ext:1"/>
-
       register :complete, :ext
 
       def reason